Update RigolInterface.py
Added a class for being able to connect the scopes in either USB or using the TCP IP protocol. Had to make few changes because an IP device can't provide a manufacturer. Kept the old class in order no to create issues in the already existing codes using the scopes